+++ tests/integration/cases/float_special_values/Cpp_float_format_fixed_v.cpp
@@ -3,8 +3,8 @@
 #include <vector>
 void check_() {
 auto my_data = std::vector<double>{
-    INFINITY,
-    -INFINITY,
-    NAN,
+    static_cast<double>(INFINITY),
+    -static_cast<double>(INFINITY),
+    static_cast<double>(NAN),
 };
 }